requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
async-generator==1.10
boto3==1.24.68
botocore==1.27.68
click==8.1.3
colorama==0.4.5
colorlover==0.3.0
csvkit==1.0.7
cufflinks==0.17.3
dbfread==2.0.7
dnspython==2.2.1
docx==0.2.4
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3
gTTS==2.2.4
h11==0.13.0
importlib-metadata==4.12.0
ipdb==0.13.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
leather==0.3.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.2
opencv-python==4.5.5.64
outcome==1.2.0
pandas==1.4.4
parsedatetime==2.4
playsound==1.3.0
protobuf==3.20.1
pymongo==4.2.0
PyMySQL==1.0.2
PyPDF2==2.0.0
python-dateutil==2.8.2
python-docx==0.8.11
pytimeparse==1.1.8
pytube==12.1.0
pytz==2022.2.1
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
SQLAlchemy==1.4.41
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
wsproto==1.2.0
zipp==3.8.1
gunicorn==20.0.4
python-dotenv~=0.21.0
certifi~=2022.6.15